added specific user check

+++ bot.py
@@ -3,17 +3,17 @@ import os
 import discord
 from discord.ext import commands
 
-TOKEN = os.getenv('DISCORD_TOKEN')
+TOKEN = os.getenv("DISCORD_TOKEN")
 
-client = commands.Bot(command_prefix='!', intents=discord.Intents.all())
+client = commands.Bot(command_prefix="!", intents=discord.Intents.all())
 
 # on startup
 
 
 @client.event
 async def on_ready():
-    print('Connected to bot: {}'.format(client.user.name))
-    print('Bot ID: {}'.format(client.user.id))
+    print("Connected to bot: {}".format(client.user.name))
+    print("Bot ID: {}".format(client.user.id))
 
 # code to mention a user whenever another user joins a voice channel
 # in this example, my friend andy will be pinged
@@ -33,12 +33,12 @@ async def on_voice_state_update(member, before, after):
 async def on_message(message):
     if not message.author.bot:
         if message.content == "pog":
-            await message.channel.send('very pog')
+            await message.channel.send("very pog")
         elif message.content == "very pog":
-            await message.channel.send('the poggest')
+            await message.channel.send("the poggest")
 
-    if len(message.attachments) > 0:
-        await message.channel.send(message.author.id)
+    if len(message.attachments) > 0 and message.author.id == 181438247015022592:
+        await message.channel.send("Nice try dumbass")
         await message.delete()
